Add weak runtime dependency on gvfs-fuse for non-Flatpak builds
This shouldn't change anything on stock installations of Fedora
Silverblue, Workstation, etc. because gvfs-fuse is already part of the
budgie-desktop, critical-path-gnome, gnome-desktop and mate-desktop
groups in fedora-comps [1], and of those gnome-desktop is part of
workstation-product-environment.  In addition, both nemo (another file
manager) [2] and totem [3] have Requires on gvfs-fuse.

Even though gvfs-fuse contains only an executable at
%{_libexecdir}/gvfsd-fuse, a configuration file and a manual page, the
dependency was tightened with the %{_isa} macro to match those on gvfs
in this package [4], and on gvfs-fuse in nemo [2,5] and totem [6].

Flatpak builds are skipped because Flatpaks use the host's GVfs and
don't need GVfs inside the container [7].

Fallout from 82b6076798be058a5ad13a2523bcc065a033304a

[1] https://pagure.io/fedora-comps

[2] Fedora nemo commit 768bdeefdb3d1d1d
    https://src.fedoraproject.org/rpms/nemo/c/768bdeefdb3d1d1d

[3] Fedora totem commit 3811c00ea0b6cea1
    https://src.fedoraproject.org/rpms/totem/c/3811c00ea0b6cea1

[4] Commit 189a33cd05d8b116
    https://src.fedoraproject.org/rpms/nautilus/c/189a33cd05d8b116

[5] Fedora nemo commit 964dec7f5b98df97
    https://src.fedoraproject.org/rpms/nemo/c/964dec7f5b98df97

[6] Fedora totem commit 9f3152117182dfba
    https://src.fedoraproject.org/rpms/totem/c/9f3152117182dfba

[7] Fedora totem commit 60349fb920e05f84
    https://src.fedoraproject.org/rpms/totem/c/60349fb920e05f84
    https://src.fedoraproject.org/rpms/totem/pull-request/3

https://bugzilla.redhat.com/show_bug.cgi?id=2352782